"Trig" calculator components moved to Calculator folder
