+++ diff-engine/src/Components/Graphs/Calculators/Trig/AngleConversion/SinCos/cos.component.jsx
