+++ diff-engine/src/Components/Graphs/Calculators/Trig/AngleConversion/SinCos/sin.component.jsx
