+++ diff-engine/src/Components/Graphs/Calculators/Trig/AngleConversion/SinCos/tan.component.jsx
