+++ diff-engine/src/Components/Graphs/Calculators/Trig/AngleConversion/TrigFunctions/function.styles.jsx
