+++ diff-engine/src/Components/Graphs/Calculators/Trig/AngleConversion/TrigFunctions/functions.component.jsx
