+++ diff-engine/src/Components/Graphs/Calculators/Trig/AngleConversion/angle_keys.component.jsx
@@ -1,5 +1,5 @@
 import { useEffect,useContext } from "react";
-import { KeyBox } from "../../../KeyPad/homekeys.styles";
+import { KeyBox } from "../../../KeyPad/input.styles";
 import { backButton,Book } from "../../../SVG";
 import { ViewContext } from "../../../../Context/view.context";
 import Button from "../../../KeyPad/Button";
