+++ diff-engine/src/Components/Graphs/Calculators/Trig/angle.data.jsx
