+++ diff-engine/src/Components/Graphs/Calculators/Trig/overlay.component.jsx
