+++ diff-engine/src/Components/Graphs/Calculators/Trig/overlay.styles.jsx
